+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Changelog Tarefas
 ======================================================
 #### 1.1.2 **/**/2024
 
-- Novo design
+Novo:
+- Faça sua anotações mais importantes
 - Suporte ao Dynamic Colors do Android 12
 - Suporte minimo Android 8
 - Suporte minimo iOS 16